0.2.1 - Moved pre-3.0 support to deprecated.php
git-svn-id: http://plugins.svn.wordpress.org/wp-bitly/trunk@260938 b8457f37-d9ea-0310-8a92-e5e31aec5664
